chore: Enable gosec and fix findings
This update involves:
- Enabling gosec in the github workflow and updating
  the github workflow versions.
- Update the codebase to fix findings from gosec. Highlights:
   - Some unhandled errors
   - Config types allowed for very large values. For example,
     the retry intervals and the commit gaps allowed for uint64.
     All values related to block sizes were left to uint64, but
     the others were reduced to uint32.
   - Prometheus server protection against Slowloris

+++ finality-provider/service/eots_manager_adapter.go
@@ -11,12 +11,12 @@ import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
 )
 
-func (fp *FinalityProviderInstance) getPubRandList(startHeight uint64, numPubRand uint64) ([]*btcec.FieldVal, error) {
+func (fp *FinalityProviderInstance) getPubRandList(startHeight uint64, numPubRand uint32) ([]*btcec.FieldVal, error) {
 	pubRandList, err := fp.em.CreateRandomnessPairList(
 		fp.btcPk.MustMarshal(),
 		fp.GetChainID(),
 		startHeight,
-		uint32(numPubRand),
+		numPubRand,
 		fp.passphrase,
 	)
 	if err != nil {
